+++ src/main/java/net/scnetwork/positioner/utils/PdfCreator.java
@@ -5,9 +5,8 @@ import com.itextpdf.text.pdf.BaseFont;
 import com.itextpdf.text.pdf.PdfPCell;
 import com.itextpdf.text.pdf.PdfPTable;
 import com.itextpdf.text.pdf.PdfWriter;
-import net.scnetwork.positioner.bean.Report;
+import net.scnetwork.positioner.domain.Report;
 
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.ArrayList;
